third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-time/timeline-usertiming-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test performance.mark/measure records on Timeline


Running: testSimplePerformanceMeasure
a

Running: testNestedPerformanceMeasure
a
b
c
d

Running: testUnbalancedPerformanceMeasure
a
b

Running: testUnnestedPerformanceMeasure
a
b

Running: testParentMeasureIsOnTop
durationTimeTotal
durationTime1
durationTime2
